refactor(agent): keep stream decoration private
